pyproject.toml
[project]
name = "opensearchfs-ruri"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
  "datasets>=4.0.0,<5.0.0",
  "einops>=0.8.0",
  "opensearch-py>=2.8.0",
  "python-dotenv>=1.0.1",
  "sentence-transformers>=3.4.0",
  "torch>=2.5.0",
  "transformers>=4.48.0",
]
